requirements.txt
discord.py==2.5.1
aiohttp==3.8.4
python-dotenv==1.0.0
pytz
